requirements.txt
azure-functions
yfinance
